requirements.txt
amqp==5.1.1
billiard==3.6.4.0
blinker==1.4
boto==2.46.1
boto3==1.13.25
botocore==1.16.25
cached-property==1.5.2
celery==5.0.2
certifi==2022.9.24
chardet==3.0.4
click==8.0.4
click-didyoumean==0.3.0
click-repl==0.2.0
configparser==5.0.1
docutils==0.15.2
idna==2.10
importlib-metadata==4.8.3
jmespath==0.10.0
kombu==5.1.0
lxml==4.9.1
mysqlclient==2.1.0
#pkg_resources==0.0.0
prompt-toolkit==3.0.32
pynamodb==4.3.3
python-dateutil==2.8.2
python-dotenv==0.20.0
python-json-logger==2.0.2
pytz==2022.6
redis==2.10.6
redis-py-cluster==1.3.6
requests==2.24.0
retrying==1.3.3
s3transfer==0.3.7
sentry-sdk==0.14.2
simplejson==3.18.0
six==1.16.0
SQLAlchemy==1.3.18
SQLAlchemy-Utils==0.36.6
typing_extensions==4.1.1
urllib3==1.25.11
vine==5.0.0
wcwidth==0.2.5
xlrd==1.1.0
xlutils==2.0.0
xlwt==1.3.0
xmltodict==0.13.0
zipp==3.6.0


git+https://thirupathi.reddy:thiru123@git.txtbox.in:8080/smcore/models.git@v4.9.8#egg=sm_models
git+https://thirupathi.reddy:thiru123@git.txtbox.in:8080/smcore/dynamo-models.git@f6dee8d4c72f953eba5dc8fc6caab7ab69bbcdb8
git+https://thirupathi.reddy:thiru123@git.txtbox.in:8080/smcore/utils.git@8af0b59d911bd30a53db4f5be06b0e333fb10375
git+https://thirupathi.reddy:thiru123@git.txtbox.in:8080/smcore/send-task.git@v3.8#egg=send_task

